+++ Sources/SwiftyHaru/Graphics/Path.swift
@@ -544,6 +544,22 @@ extension Path._PathConstructionOperation: Equatable {
     }
 }
 
+extension Path {
+    
+    /// Rules for determining which regions are interior to a path, used by the `DrawingContext.clip(to:rule:_:)`
+    /// and `DrawingContext.fill(_:rule:stroke:`) methods.
+    public enum FillRule {
+        
+        /// A rule that considers a region to be interior to a path
+        /// based on the number of times it is enclosed by path elements.
+        case evenOdd
+        
+        /// A rule that considers a region to be interior to a path
+        /// if the winding number for that region is nonzero.
+        case winding
+    }
+}
+
 extension Path: Equatable {
     
     /// Returns a Boolean value indicating whether two values are equal.